clientlens-crud-starter/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymssql==2.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.4.0
